+++ app/bot/index.js
@@ -3,6 +3,9 @@
 const debug = require('debug')('bot/index.js');
 
 exports.register = function (server, options, next) {
+    //console.log(__dirname);
+    require('local-modules')('app/bot');
+
     // instance of the telegram bot
     const bot = require('./bot')(server, options);
     server.decorate('server', 'bot', bot);
